First commit
This is the repository's first commit.

I have not included the compiled .scr file because the code contains a bug that prevents the that duplicates the screensaver and if your lucky end the process before running out of RAM.

You have been warned

But hey if you. are a fellow G.l.a.D.O.S   who loves to use their PC as a "Lab Rat" for science, Go ahead and warm up the neurotoxin emitters.

Neurotoxin not included and the cake is a lie.

Ok enough Portal jokes,

Have a great day.